+++ examples/react-typescript-tailwindcss/esbuild.renderer.config.js
@@ -3,12 +3,8 @@ const postCss = require('esbuild-plugin-postcss2').default
 const autoprefixer = require('autoprefixer')
 const tailwindcss = require('tailwindcss')
 
-/**
- * @param {Partial<import('esbuild').BuildOptions>} merge
- *
- * @return {import('esbuild').BuildOptions}
- */
-module.exports = (merge) => ({
+/** @var {Partial<import('esbuild').BuildOptions>} */
+module.exports = {
   platform: 'browser',
   entryPoints: [path.resolve('src/renderer/index.tsx')],
   bundle: true,
@@ -22,6 +18,5 @@ module.exports = (merge) => ({
     postCss({
       plugins: [autoprefixer, tailwindcss],
     }),
-  ],
-  ...merge,
-})
+  ]
+}