+++ examples/react-typescript/esbuild.main.config.js
@@ -1,17 +1,12 @@
 const path = require('path')
 
-/**
- * @param {Partial<import('esbuild').BuildOptions>} merge
- *
- * @return {import('esbuild').BuildOptions}
- */
-module.exports = (merge) => ({
+/** @var {Partial<import('esbuild').BuildOptions>} */
+module.exports = {
   platform: 'node',
   entryPoints: [path.resolve('src/main/main.ts')],
   bundle: true,
   target: 'node14.16.0', // electron version target
   loader: {
     '.ts': 'ts',
-  },
-  ...merge,
-})
+  }
+}