+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+## v1.2.0
+
+## Bug fixes
+
+- NODE_ENV was not set in the user configuration
+
+## Breaking changes
+
+**Users configuration file is no longer a function**
+
+You need to export an object from your esbuild configuration or your webpack configuration
+
+From
+
+```javascript
+module.exports = (merge) => ({
+  ...yourConfig,
+  ...merge,
+})
+```
+
+To
+
+```javascript
+module.exports = {
+  ...yourConfig,
+}
+```
+
 ## v1.1.6
 
 ### Bug fixes